Drop commented out XMLPath record from HDF5 plugin
After b9b26f02 (Improved XML error handling and fixed bug on file
close., 2014-01-24), we no longer need the XMLPath record, as
XMLFileName supports the whole path. As there is no future need for this
record, simply drop this dead record from the database entirely.

Because this is version controlled, there is always the possibility of
taking it back if needed.