Document HDF5 dataset fill value record
This record is available since 7a0d0176 (Added fillValue parameter to
the HDF5 file writer., 2015-09-22), but it was never properly
documented.

Document it under the metadata section as it is basically a metadata
regarding the default value the readers should take when no value is
directly. This has been documented based on the corresponding HDF5
library function call that uses this value.

[1] https://support.hdfgroup.org/documentation/hdf5/latest/group___d_c_p_l.html#ga4335bb45b35386daa837b4ff1b9cd4a4